% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charname.R
\name{charname}
\alias{charname}
\title{Categoric column name}
\usage{
charname(x = "Data frame")
}
\arguments{
\item{x}{A data frame}
}
\value{
A character vector.
}
\description{
Categoric column name
}
\examples{
cat<-data.frame(col1=c("cat1","cat2","cat3","cat2","cat1"),col2=c(1,2,3,4,5))
charname(cat)
}
